Added temp routes and fixed bugs
